Fix a compile warning and maybe a run-time error (see OtherBill's comment in time/win32/time.c on another call to Sleep)
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@61413 13f79535-47bb-0310-9956-ffa450edef68

+++ network_io/win32/poll.c
@@ -119,7 +119,7 @@ APR_DECLARE(apr_status_t) apr_poll(apr_pollfd_t *aprset, apr_int32_t *nsds,
     }
 
     if (newread == NULL && newwrite == NULL && newexcept == NULL) {
-        Sleep(timeout / 1000); /* convert microseconds into milliseconds */
+        Sleep((DWORD)(timeout / 1000)); /* convert microseconds into milliseconds */
         return APR_TIMEUP; /* TODO - get everybody in synch with Win32 apr_status_t */
     }
     else {
